dev-requirements.txt
aioresponses
asynctest
flake8
mock
pytest
pytest-aiohttp
pytest-cache
pytest-capturelog
pytest-cover
pytest-sugar
swagger-spec-validator
tox
wheel
zest.releaser
